Sequoia v1.1

+++ SequoiaApp/views.py
@@ -5,7 +5,7 @@ from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
 from django.core import serializers
 from django.http import JsonResponse
-from django.shortcuts import get_object_or_404, HttpResponseRedirect, redirect, render, render_to_response
+from django.shortcuts import get_object_or_404, HttpResponseRedirect, redirect, render, render_to_response, Http404
 from django.utils.decorators import method_decorator
 from django.views.generic import UpdateView
 from django.forms.models import model_to_dict
@@ -41,32 +41,54 @@ def remitter_create(request):
                 account_number=form.cleaned_data.get('account_number'),
                 mobile_number=form.cleaned_data.get('mobile_number'),
                 PAN=form.cleaned_data.get('PAN'),
-                GSTIN=form.cleaned_data.get('GSTIN')
+                GSTIN=form.cleaned_data.get('GSTIN'),
+                created_by=request.user
             )
-        return HttpResponseRedirect(reverse(remitter_search))
+            return HttpResponseRedirect(reverse(remitter_search))
     return render(request, 'SequoiaApp/remitter_create.html', {'form': form})
 
 
 @login_required
 def remitter_search(request):
     """This function is used to search Remitters"""
-    remitter_list = Remitter.objects.all()
+    user = request.user
+    if user.is_superuser:
+        remitter_list = Remitter.objects.all()
+    else:
+        remitter_list = Remitter.objects.filter(created_by=user)
     return render(request, 'SequoiaApp/remitter_search.html', {'filter': remitter_list})
 
 
 @method_decorator(login_required, name='dispatch')
 class RemitterUpdateView(UpdateView):
     """This function is used for Updating Remitter Data"""
-    fields = ('name', 'account_number', 'mobile_number', 'PAN')
+    fields = ('name', 'account_number', 'mobile_number', 'PAN', 'GSTIN')
     model = Remitter
+
+    def user_passes_test(self, request):
+        if request.user.is_authenticated:
+            self.object = self.get_object()
+            return self.object.created_by == request.user
+        else:
+            return False
+
+    def dispatch(self, request, *args, **kwargs):
+        if request.user.is_superuser:
+            return super(RemitterUpdateView, self).dispatch(request, *args, **kwargs)
+        elif not self.user_passes_test(request):
+            return redirect("remitter_search")
+        return super(RemitterUpdateView, self).dispatch(request, *args, **kwargs)
 
 
 @login_required
 def remitter_delete(request, pk=None):
     """This function is used to delete Remitter"""
     remitter = get_object_or_404(Remitter, pk=pk)
-    remitter.delete()
-    return redirect("remitter_search")
+    if request.user == remitter.created_by or request.user.is_superuser:
+        remitter.delete()
+        return redirect("remitter_search")
+    else:
+        return redirect("remitter_search")
 
 
 @login_required
@@ -84,7 +106,8 @@ def customer_create(request):
                 bank_ifsc_code=form.cleaned_data.get('bank_ifsc_code'),
                 PAN=form.cleaned_data.get('PAN'),
                 mobile_number=form.cleaned_data.get('mobile_number'),
-                GSTIN=form.cleaned_data.get('GSTIN')
+                GSTIN=form.cleaned_data.get('GSTIN'),
+                created_by=request.user
             )
             return HttpResponseRedirect(reverse(customer_search))
     return render(request, 'SequoiaApp/customer_create.html', {'form': form})
@@ -93,7 +116,11 @@ def customer_create(request):
 @login_required
 def customer_search(request):
     """This function is used to search Customers"""
-    customer_list = Customer.objects.all()
+    user = request.user
+    if user.is_superuser:
+        customer_list = Customer.objects.all()
+    else:
+        customer_list = Customer.objects.filter(created_by=user)
     return render(request, 'SequoiaApp/customer_search.html', {'filter': customer_list})
 
 
@@ -104,27 +131,43 @@ class CustomerUpdateView(UpdateView):
               'bank_ifsc_code', 'PAN', 'mobile_number', 'GSTIN')
     model = Customer
 
+    def user_passes_test(self, request):
+        if request.user.is_authenticated:
+            self.object = self.get_object()
+            return self.object.created_by == request.user
+        else:
+            return False
+
+    def dispatch(self, request, *args, **kwargs):
+        if request.user.is_superuser:
+            return super(CustomerUpdateView, self).dispatch(request, *args, **kwargs)
+        elif not self.user_passes_test(request):
+            return redirect("remitter_search")
+        return super(CustomerUpdateView, self).dispatch(request, *args, **kwargs)
+
 
 @login_required
 def customer_delete(request, pk=None):
     """This function is used to delete Customers"""
     customer = get_object_or_404(Customer, pk=pk)
-    customer.delete()
-    return redirect("customer_search")
+    if request.user == customer.created_by or request.user.is_superuser:
+        customer.delete()
+        return redirect("customer_search")
+    else:
+        return redirect("customer_search")
 
 
 @login_required
 def rtgs_create(request):
     # get or create the form
     form = RTGSForm(request.POST or None)
-
     if form.is_valid():
         # Get the customer id (hidden field) from the form
         customer_id = form.cleaned_data.get('customer_id')
         try:
             # get the customer instance
             customer = Customer.objects.get(id=customer_id)
-        except Customer.DoesNotExist:
+        except (Customer.DoesNotExist, ValueError):
             # else create a new one
             customer = Customer()
 
@@ -135,7 +178,7 @@ def rtgs_create(request):
                     setattr(customer, key, value)
         
         # finally
-        print("customer: {}".format(customer))
+        customer.created_by = request.user
         customer.save()
 
         # use the customer instance created above
@@ -144,15 +187,16 @@ def rtgs_create(request):
             remitter=form.cleaned_data.get('remitter'),
             cheque_number=form.cleaned_data.get('cheque_number'),
             amount_in_figure=form.cleaned_data.get('amount_in_figure'),
-            amount_in_word=form.cleaned_data.get('amount_in_word')
+            amount_in_word=form.cleaned_data.get('amount_in_word'),
+            created_by=request.user
         )
 
         # redirect to the RTGS view, so that when the user navigates back
         # the browser doesn't POST again
         return redirect("rtgs", pk=rtgs.pk)
     
-    # fetch all the remitters to display as dropdown in the RTGS create form
-    remitters = Remitter.objects.all()
+    # fetch the remitters created by the current logged in user to display as dropdown in the RTGS create form
+    remitters = Remitter.objects.filter(created_by=request.user)
 
     # also: activate the "Create RTGS" in the side bar
     return render(request, 'SequoiaApp/rtgs_create.html', {'form': form, 'remitters': remitters, 'active_create_rtgs': 'active'})
@@ -164,15 +208,19 @@ def rtgs_form(request):
 
 
 @login_required
-def rtgs_list(request):
-    rtgs_list = RTGS.objects.all()
-    return render(request, 'SequoiaApp/rtgs_list.html', {'rtgs': rtgs_list})
+def rtgs_search(request):
+    user = request.user
+    rtgs_list = RTGS.objects.filter(created_by=user)
+    return render(request, 'SequoiaApp/rtgs_list.html', {'rtgs_list': rtgs_list})
 
 
 @login_required
 def rtgs(request, pk):
     rtgs = get_object_or_404(RTGS, pk=pk)
-    return render(request, 'SequoiaApp/rtgs_form.html', {'rtgs': rtgs})
+    if request.user == rtgs.created_by or request.user.is_superuser:
+        return render(request, 'SequoiaApp/rtgs_form.html', {'rtgs': rtgs})
+    else:
+        return redirect("rtgs_search")
 
 
 @login_required
